[aon_timer] Support 64-bit counter registers on earlgrey_1.0.0

+++ tools/tock-configurator/generator/src/tock_main.rs
@@ -270,7 +270,7 @@ impl<C: Chip + 'static> TockMain<C> {
                 const NAME: &'static str = "qemu";
                 const CPU_FREQ: u32 = 24_000_000;
                 const PERIPHERAL_FREQ: u32 = 6_000_000;
-                const AON_TIMER_FREQ: u32 = 250_000;
+                const AON_TIMER_FREQ: u64 = 250_000;
                 const UART_BAUDRATE: u32 = 115200;
             }
 
@@ -279,7 +279,7 @@ impl<C: Chip + 'static> TockMain<C> {
                 const NAME: &'static str = "fpga";
                 const CPU_FREQ: u32 = 24_000_000;
                 const PERIPHERAL_FREQ: u32 = 6_000_000;
-                const AON_TIMER_FREQ: u32 = 250_000;
+                const AON_TIMER_FREQ: u64 = 250_000;
                 const UART_BAUDRATE: u32 = 115200;
             }
 
@@ -288,7 +288,7 @@ impl<C: Chip + 'static> TockMain<C> {
                 const NAME: &'static str = "silicon";
                 const CPU_FREQ: u32 = 100_000_000;
                 const PERIPHERAL_FREQ: u32 = 24_000_000;
-                const AON_TIMER_FREQ: u32 = 200_000;
+                const AON_TIMER_FREQ: u64 = 200_000;
                 const UART_BAUDRATE: u32 = 115200;
             }
 
@@ -299,7 +299,7 @@ impl<C: Chip + 'static> TockMain<C> {
                 // Clock frequencies as of https://github.com/lowRISC/opentitan/pull/19368
                 const CPU_FREQ: u32 = 500_000;
                 const PERIPHERAL_FREQ: u32 = 125_000;
-                const AON_TIMER_FREQ: u32 = 125_000;
+                const AON_TIMER_FREQ: u64 = 125_000;
                 const UART_BAUDRATE: u32 = 7200;
             }
